Initialize unspecified fields of literal struct with zero.
This bug has caused an issue that are very hard to find what was wrong,
because uninitialized fields get garbages from the stack and the
resulting issues occur far from the point at where we allocate literal
structs. In this case, the source code with this bug produce a binary
that can pass all unit tests which cannot produce the second generation
binary that can pass unit tests.

It took me three days to find this bug.