Run unit tests with the system's default CC.
These tests were compiled only with 8cc, so we were not sure if
they really make sense.

This patch adds a new make target "test2" to compile and run the
tests with the system's default C compiler. That target is run for
"fulltest" target too.

I found many tests succeed (or failed) only with 8cc.
They are now #ifdef'ed out. We need to revisit them to fix later.