Add #pragma once.
Real compliers are more clever than 8cc on this. They use some
heuristics to determine if two files are the same, such as checking
the last modified time of the two.

8cc doesn't do that, so if pathes are different, the same file
can be included more than once even if the file contains a #pragma
once directive. That needs to be fixed, but we've got more thing
to do before that.