Create default output file in the current directory.
8cc used to create the output file in the same directory as the source
file if no -o option is given. That's an unusual behavior as a C compiler,
because other C compliers create the default output file in the current
directory.

8cc now follows the other compilers. If no -o option is given, the output
file name is created by stripping all directory components from the input
path and replacing the suffix accordingly (.c -> .o or .s).