Cleanup: remove gettime().
Previously we call time(2) when we need a timestamp and cache the result.

This patch makes the initializer call time(2) unconditionally
and set the result to a global variable. This simplifies the
code a bit.

The cost of time(2) is negligible. I guess it doesn't even enter
to the kernel because it can be implemented using the user-accessible
TSC counter.

The reason why we store the result to the global variable instead of
calling time(2) every time we need a timestamp is because we want
consistent results. I don't want 8cc to expand __TIME__ macros to
different strings when it processes a long source code which takes
more than 1 second to finish reading.