Do not use realpath() for #include.
I don't want to call realpath() for each #include because
the function invokes too many system calls. (It calls stat()
on each path component.)

This patch is to uniquify paths only with textual processing.
The result would be less accurate than realpath() because it
couldn't uniquify symbolic links, but the new function should
be good enough.