Rewrite character input stream.
The previous implementation of File depended on FILE*.
When we want to read from a string, we needed to write the string
to a temporary file to read back.

I realized that there are many ocasssions we want to read
from a string. One example is ##. A string created by token
concatenation needs to be tokenized as if it were read from a file.
We didn't handle that well.

With the new implementation of File, reading from string is
as easy as reading from file.