Fix string literals containing null bytes.
We didn't allow null bytes in string literals. "a\0" was interpreted
as the same as "a". Obviously this is wrong. String literals should be
able to contain any bytes including NUL. This patch fixes the bug.

That means the data type for string literals now needs to have both
data and size fields. I added "slen" field to the Token type for that
purpose.