Fix _Alignas used at top level.
It's important for the parser to distinguish function definitions from
declarations when reading top-level construsts. Unfortunately, the C
grammar made it hard. Currently, we reads-ahead arbitrary number of
tokens until we see a hint about whether what we are reading is
a function definition or not.

The previous implementation of the function was confused when it sees
an _Alignas() declaration specifier, because it didn't expect any
specifier having parentheses as part of it. This patch fixes that bug.