Support primitive address constant expressions.
C11 has notion of address constants, which are addresses of global
variables or functions. You can use an address constant + an integer
constant expression as an rvalue in initializers.

The implementation of 8cc's integer constant expression is incomplete.
This patch is also incomplete or even buggy. I think it's better than
nothing, though.

8cc is now able to compile

  int x[10];
  int *p = &x + 3;