Remove redundant "void" from function parameter list.
If a function doesn't have any parameter, we should put "void"
to the parameter list to tell the compiler that the parameter list
is actually empty instead of being omitted in K&R-style.

But that's only meaningful for function declarations. For function
definitions, "void" and the empty parameter list are the same.
See C11 6.7.5.3p14. It's defined like that since C89.
I honestly misunderstood that until now.

What's funny is that I implemented the correct behavior in 8cc.
I must have read that sentence, understood that, and implemented that.
But I didn't realize what the sentence actually means.